seperate Testclass for Moving Cards + Setup Method
